requirements.txt
attrs==19.3.0
blis==0.4.1
catalogue==0.0.8
certifi==2019.11.28
chardet==3.0.4
cymem==2.0.3
Django==2.2
djangorestframework==3.10.3
docx2txt==0.8
en-core-web-sm==2.2.5
idna==2.8
importlib-metadata==1.3.0
jsonschema==3.2.0
more-itertools==8.0.2
murmurhash==1.0.2
nltk==3.4.5
numpy==1.17.4
pandas==0.25.3
pdfminer.six==20191110
plac==1.1.3
preshed==3.0.2
pycparser==2.19
pycryptodome==3.9.4
pyresparser==1.0.5
pyrsistent==0.15.6
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
six==1.13.0
sortedcontainers==2.1.0
spacy==2.2.3
sqlparse==0.3.0
srsly==0.2.0
thinc==7.3.1
tqdm==4.40.2
urllib3==1.25.7
wasabi==0.4.2
zipp==0.6.0
python -m spacy download en_core_web_sm
python -m nltk.downloader words
